Addressing PR comments; splitting out routing e2e tests.
